+++ Tests/VCoreMacrosTests/URLMacroTests.swift
@@ -5,10 +5,14 @@
 //  Created by Vakhtang Kontridze on 07.01.24.
 //
 
+#if canImport(VCoreMacrosImplementation)
+
+import Foundation
 import XCTest
 import SwiftSyntax
 import SwiftSyntaxMacros
 import SwiftSyntaxMacrosTestSupport
+@testable import VCoreMacros
 @testable import VCoreMacrosImplementation
 
 // MARK: - Tests
@@ -64,3 +68,5 @@ final class URLMacroTests: XCTestCase {
         )
     }
 }
+
+#endif